requirements.txt
langchain
langchain-community
langchain-openai
openai
beautifulsoup4
requests
tqdm
python-dotenv
lxml
pytesseract
pymupdf